myprojectDj/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
requests==2.32.0
python-telegram-bot==22.1